add separate `key`-argument to `SetCommand` (resolves #13)
