+++ src/command/idb-command/idb-update-command.ts
@@ -1,10 +1,9 @@
 import { isNull, MissingKeyError, NdbDocument, NotFoundError } from '@normalized-db/core';
-import { ObjectStore } from 'idb';
 import { EmptyInputError } from '../../error/empty-input-error';
 import { UpdateCommand } from '../update-command';
-import { IdbBaseWriteCommand } from './idb-base-write-command';
+import { IdbBaseUpdateCommand } from './idb-base-update-command';
 
-export class IdbUpdateCommand<T extends NdbDocument> extends IdbBaseWriteCommand<T> implements UpdateCommand<T> {
+export class IdbUpdateCommand<T extends NdbDocument> extends IdbBaseUpdateCommand<T> implements UpdateCommand<T> {
 
   /**
    * @inheritDoc
@@ -21,24 +20,5 @@ export class IdbUpdateCommand<T extends NdbDocument> extends IdbBaseWriteCommand
     }
 
     return this.executeHelper(data, isPartialUpdate);
-  }
-
-  protected async executeHelper(data: T | T[], isPartialUpdate = false): Promise<boolean> {
-    const objectStore = (await this._context.read(this._type)).objectStore(this._type);
-    if (Array.isArray(data)) {
-      await Promise.all(data.map(item => this.checkExistence(objectStore, item)));
-    } else {
-      await this.checkExistence(objectStore, data);
-    }
-
-    return super.write(data, null, isPartialUpdate);
-  }
-
-  private async checkExistence(objectStore: ObjectStore, item: T): Promise<void> {
-    const key = this.getKey(item);
-    const existingItem = objectStore.get(key);
-    if (isNull(existingItem)) {
-      throw new NotFoundError(this._type, key);
-    }
   }
 }
