+++ src/command/idb-command/index.ts
@@ -1,4 +1,5 @@
 import { IdbBaseCommand } from './idb-base-command';
+import { IdbBaseUpdateCommand } from './idb-base-update-command';
 import { IdbBaseWriteCommand } from './idb-base-write-command';
 import { IdbClearCommand } from './idb-clear-command';
 import { IdbCommandFactory } from './idb-command-factory';
@@ -12,6 +13,7 @@ export {
   IdbBaseCommand,
   IdbBaseWriteCommand,
   IdbCreateCommand,
+  IdbBaseUpdateCommand,
   IdbUpdateCommand,
   IdbPutCommand,
   IdbRemoveCommand,
